restructure evaluation tests
